lab1/training_service/requirements.txt
pandas
numpy
scikit-learn
joblib
google-cloud-storage 